requirements.txt
# For OpenCV character recognition
numpy==2.2.6
opencv-python==4.12.0.88
pdf2image==1.17.0
pdfplumber==0.11.7
Pillow==12.0.0
pytesseract==0.3.13
